♻️ Lead capture: extract the multi-step form into a reusable component
Pull the 3-screen form card (progress, validation, /api/lead-capture
submit) out of the V3LeadCapture block into LeadCaptureForm so it can
be reused outside the block. The success line becomes a prop; the
/consulting/react block passes its React-specific copy, so that page
is unchanged.
